+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Complete rewrite of whole wrestic
+- Better logging
+- Better code design
+- Better error handling
+- Better maintainability
+- Less deadlock potential with command IOs
 
 ## [v0.1.9] 2020-05-04
 ### Changed