+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Make waiting for locks more stable
 - Don't drop k8sExe stderr output ([#27])
 
+
 ## [v0.2.0] 2020-05-29
 ### Changed
 - Complete rewrite of whole wrestic