+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.2.1] 2021-01-07
 ### Fixed
 - Make waiting for locks more stable ([#24])
 - Don't drop k8sExe stderr output ([#27])
@@ -134,6 +134,7 @@ compatibility with older operator versions. Changes to the design contain:
 
 [Unreleased]: https://git.vshn.net/vshn/wrestic/compare/v0.2.1...master
 
+[v0.2.1]: https://github.com/vshn/wrestic/releases/tag/v0.2.1
 [v0.2.0]: https://github.com/vshn/wrestic/releases/tag/v0.2.0
 [v0.1.9]: https://github.com/vshn/wrestic/releases/tag/v0.1.9
 [v0.1.8]: https://github.com/vshn/wrestic/releases/tag/v0.1.8